requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
celery[redis]==5.3.4
redis==4.6.0
httpx==0.25.1
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
selectolax==0.3.17
lxml==4.9.3
openai==1.3.5
anthropic==0.7.7
langdetect==1.0.9
qdrant-client==1.6.9
numpy==1.25.2
scikit-learn==1.3.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.0
sentence-transformers==3.3.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.1.2
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
python-dateutil==2.8.2
aiosmtplib==3.0.0
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6
gunicorn==21.2.0
yfinance==0.2.28
alpha-vantage==2.3.1
pytz==2023.3

# Scheduler and async support
apscheduler==3.10.4
websockets==12.0

# Additional dependencies for advanced financial AI
tweepy==4.14.0
praw==7.7.1
discord.py==2.3.2
pandas==2.0.3

# Deep Learning (optional - commented out for lighter deployment)
# tensorflow==2.13.0
# keras==2.13.1
